cms/api/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.4
pydantic-settings==2.3.4
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.9
asyncpg==0.29.0
